ong-adjahi-platform/backend/requirements.txt
# Backend Dependencies - ONG ADJAHI Platform

# Django Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-otp==1.3.0
qrcode==7.4.2
pyotp==2.9.0
cryptography==42.0.0

# Celery & Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# API Documentation
drf-spectacular==0.27.0

# File Storage
boto3==1.34.19
django-storages==1.14.2

# Image Processing
Pillow==10.2.0

# Excel/PDF Generation
openpyxl==3.1.2
reportlab==4.0.9
WeasyPrint==60.2

# SMS/Email
twilio==8.11.1
sendgrid==6.11.0

# Monitoring & Logging
sentry-sdk==1.39.2
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
pylint==3.0.3
mypy==1.8.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
